requirements.txt
httpx==0.25.2
numpy==1.26.2
pillow==10.1.0